+++ packages/flutter_scene_layout3d/lib/src/sliver/sliver_grid.dart
@@ -5,6 +5,7 @@ import '../geometry/constraints3d.dart';
 import '../geometry/offset3d.dart';
 import '../built_children.dart';
 import '../layout3d.dart';
+import '../layout_pass.dart';
 import '../scroll/grid_view.dart' show Grid3dDelegate, Grid3dLayout;
 import 'sliver.dart';
 import 'sliver_constraints.dart';
@@ -19,6 +20,7 @@ import 'sliver_constraints.dart';
 class SliverGrid3d extends Sliver3d
     with
         Layout3dWithChildrenMixin<ParentData3d>,
+        Layout3dLayoutPassMixin,
         Layout3dBuiltChildrenMixin<ParentData3d> {
   /// Creates a grid over an explicit set of children.
   SliverGrid3d({